# backend/docs/swagger/openapi.yml
openapi: 3.0.3
info:
  title: ERP Distribution - API Gateway
  version: 1.0.0
  description: |
    Specification OpenAPI centralisee exposee via l'API Gateway.
    Les clients consomment principalement ces routes /api/v1/*.
servers:
  - url: http://localhost:4000
    description: API Gateway locale
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: erp_session
paths:
...
  /api/v1/suppliers:
    get:
      summary: Lister les fournisseurs (via gateway)
      tags: [Suppliers]
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        '200':
          description: Liste fournisseurs
        '403':
          description: Permission suppliers.read requise
    post:
      summary: Creer un fournisseur (via gateway)
      tags: [Suppliers]
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                supplier_id:
                  type: string
                  description: Optionnel (auto-genere si absent, ex SUP-051)
                supplier_name:
                  type: string
                country:
                  type: string
                contact_email:
                  type: string
              required: [supplier_name]
      responses:
        '201':
          description: Fournisseur cree
        '403':
          description: Permission suppliers.create requise
  /api/v1/suppliers/{supplierId}:
    get:
      summary: Detail fournisseur (via gateway)
      tags: [Suppliers]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: supplierId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detail fournisseur
        '404':
          description: Non trouve
        '403':
          description: Permission suppliers.read requise
    patch:
      summary: Mettre a jour un fournisseur (via gateway)
      tags: [Suppliers]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: supplierId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fournisseur mis a jour
        '403':
          description: Permission suppliers.update requise
    delete:
      summary: Supprimer un fournisseur (via gateway)
      tags: [Suppliers]
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: supplierId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fournisseur supprime
        '403':
          description: Permission suppliers.delete requise
...